+++ src/main/java/module/series/SeriesModule.java
@@ -32,5 +32,4 @@ public final class SeriesModule extends DefaultModule {
 	public KeyStroke getKeyStroke() {
 		return KeyStroke.getKeyStroke(KeyEvent.VK_F3, 0);
 	}
-	
 }